# docs/swagger.yaml
definitions:
...
  controller.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  controller.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    type: object
...
  models.User:
    properties:
      biography:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_blocked:
        type: boolean
      job_function:
        type: string
      keycloak_id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      phone_number_visible:
        type: boolean
      profile_photo_url:
        type: string
      sector:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: If the email exists, a reset message will be sent.
      parameters:
      - description: Forgot password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request a password reset
      tags:
      - Auth
...
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the provided token and new password.
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password using a token
      tags:
      - Auth
...
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user in the database and Keycloak
      parameters:
      - description: User info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Register a new user
      tags:
      - Users
...
swagger: "2.0"